# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts/{contactId}/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs for a contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Add a job to a contact
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /contacts/{contactId}/jobs/{id}:
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        serviceType:
          type: string
        price:
          type: number
        date:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - serviceType
        - price
        - createdAt
    CreateJobBody:
      type: object
      properties:
        serviceType:
          type: string
        price:
          type: number
        date:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
      required:
        - serviceType
        - price
    UpdateJobBody:
      type: object
      properties:
        serviceType:
          type: string
        price:
          type: number
        date:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
...